Add blas lapack functions.

+++ inc/tensor_hao.h
@@ -111,7 +111,7 @@ namespace tensor_hao
      //=========
      //FUNCTIONS
      //=========
-     Tensor_hao_ref<T, D-1> operator[] (size_t i)
+     Tensor_hao_ref<T, D-1> operator[] (int i)
      {
          if( i > ( this->n[D-1] ) || i<0 )
          {
